+++ packages/terra-site/src/examples/section-header/Index.jsx
@@ -2,8 +2,6 @@
 import React from 'react';
 import PropsTable from 'terra-props-table';
 import Markdown from 'terra-markdown';
-import IndexTemplate from 'terra-site/src/IndexPageTemplate';
-
 import ReadMe from 'terra-section-header/docs/README.md';
 import { version } from 'terra-section-header/package.json';
 
@@ -13,21 +11,32 @@ import SectionHeaderSrc from '!raw-loader!terra-section-header/src/SectionHeader
 
 // Example Files
 import DefaultSectionHeader from './DefaultSectionHeader';
-import DefaultSectionHeaderSrc from '!raw-loader!./DefaultSectionHeader.jsx';
+import LongTitleSectionHeader from './LongTitleSectionHeader';
+import LongTitleAccordionSectionHeader from './LongTitleAccordionSectionHeader';
+import ClosedSectionHeader from './ClosedSectionHeader';
+import OpenSectionHeader from './OpenSectionHeader';
+import OnClickSectionHeader from './OnClickSectionHeader';
+import AccordionSectionHeader from './AccordionSectionHeader';
+
+const Examples = () => (
+  <div>
+    <h2>Examples</h2>
+    <DefaultSectionHeader />
+    <LongTitleSectionHeader />
+    <LongTitleAccordionSectionHeader />
+    <ClosedSectionHeader />
+    <OpenSectionHeader />
+    <OnClickSectionHeader />
+    <AccordionSectionHeader />
+  </div>
+);
 
 const SectionHeaderExamples = () => (
   <div>
     <div id="version">Version: {version}</div>
     <Markdown id="readme" src={ReadMe} />
-
-    <h1 style={{ paddingBottom: '0.3em', borderBottom: '1px solid #eaecef' }}>Examples</h1>
-    <IndexTemplate
-      title="SectionHeader - Default"
-      example={<DefaultSectionHeader />}
-      exampleSrc={DefaultSectionHeaderSrc}
-    />
-
     <PropsTable id="props" src={SectionHeaderSrc} />
+    <Examples />
   </div>
 );